D20 with stats
Just a fun learning project. I play DnD, and I wanted to make a random D20 roller, but i wanted to also see how random it was, The D20 roller was easy, most of the rest of the code was figuring out how to track the stats of how random it was.